bulkemail/requirements.txt
asgiref==3.2.10
Django==3.1
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
xlrd==1.2.0
